refactor(models): consolidate PermissionSubscriber into models/admin and drop CHANGELOG
- Move PermissionSubscriber from models/permissions.py into models/admin.py,
  co-locating all admin-domain models in one file
- Delete models/permissions.py
- Update import sites in modules/admin.py (merged into single import line)
  and bot.py (deferred import in on_ready)
- Remove legacy CHANGELOG (raw git-log dump, superseded by git history)
